requirements.txt
pymongo
requests
python-dotenv
sentence-transformers
torch
transformers